+++ scripts/legacyMacros.js
@@ -373,6 +373,7 @@ export {blessedHealer} from './macros/2014/classFeatures/cleric/lifeDomain/bless
 export {supremeHealing} from './macros/2014/classFeatures/cleric/lifeDomain/supremeHealing.js';
 export {colossusSlayer} from './macros/2014/classFeatures/ranger/hunter/colossusSlayer.js';
 export {lanceOfLethargy} from './macros/2014/classFeatures/warlock/invocations/lanceOfLethargy.js';
+export {vengefulAncestors} from './macros/2014/classFeatures/barbarian/ancestralGuardian/vengefulAncestors.js';
 // Race Features
 export {aasimarNecroticShroud, aasimarRadiantConsumption, aasimarRadiantSoul, celestialRevelation} from './macros/2014/raceFeatures/aasimar/celestialRevelation.js';
 export {acidBreathWeapon, coldBreathWeapon, fireBreathWeapon, lightningBreathWeapon, poisonBreathWeapon} from './macros/2014/raceFeatures/dragonborn/chromatic/breathWeapon.js';