+++ scripts/macros/2014/classFeatures/barbarian/ancestralGuardian/vengefulAncestors.js
@@ -1,0 +1,5 @@
+export let vengefulAncestors = {
+    name: 'Vengeful Ancestors',
+    version: '1.2.6',
+    rules: 'legacy'
+};